backend/requirements.txt
Django==2.2.4
django-environ==0.4.5
django-extensions==2.2.1
Pillow==6.1.0
psycopg2-binary==2.8.3
requests==2.22.0
Werkzeug==0.15.5
raven==6.10.0
djangorestframework==3.10.2
